[Refactor] Use constant for environment variable enabled values
This change replaces the hardcoded tuple of strings used to check for enabled environment variable values with a module-level constant '_ENABLED_VALUES'. This improves readability and maintainability by centralizing the definition of these values, making it easier to understand and update if needed in the future.

The logic and behavior of the code remain unchanged, as only the way the enabled values are defined and referenced has been altered. No functionality is affected.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I643d10fc27aa6e668812a3bdb40e8a22213cafee