[Refactor] Use os.getenv in RolloutControlledFeature._check_env_control_flag
This change refactors the _check_env_control_flag method to use os.getenv instead of checking os.environ directly.
This improves readability and makes the code more concise by leveraging the built-in functionality of os.getenv to handle cases where the environment variable is not set.
The logic and behavior of the method remain unchanged, as os.getenv returns None when the variable is not found, which is consistent with the original implementation.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ifaaaaa28cf7812e54f730c2807e8414e7aaf34e9