[Refactor] Improve assertions in result_reporter_unittest.py
This change replaces the usage of `.get()` followed by `assertIsNotNone()`
with `assertIn()` and direct dictionary access. This improves readability
and provides better error messages when the key is missing in the dictionary.
Instead of failing with "IsNotNone failed", it will fail with "X not in Y",
which is more informative.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I56837add50b97dd1cb5ff5d20e9d9d8e1039e4de