[Fix] [Refactor] Fix copy-paste error in result_reporter_unittest
In 'test_update_stats_passed', the assertion for 'ignored' count was checking the old 'group' object instead of the new 'group2' object. This change fixes the assertion to verify the correct object's state. This does not change the logic of the production code, only improves the test correctness.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ie5eb89e701a9669ef0c89677757638ae55a843c8