[Refactor] Use autospec=True in rollout_control_unittest
This change adds `autospec=True` to the `mock.patch` decorators for `hashlib.sha256` in `rollout_control_unittest.py`.

Using `autospec=True` ensures that the mock object adheres to the interface of the object being mocked. This prevents tests from passing if the mock is used incorrectly (e.g., calling non-existent methods or with incorrect arguments) or if the mocked API changes in a way that breaks compatibility. This improves the robustness of the unit tests without altering the existing logic or functionality of the code being tested.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ic47f402aa9ae07f310eed2de9f72843b884f3a7f