[Refactor] Move comment for clarity and use local var in setUp
Moved the comment explaining the mock hash value to the definition of the constant for better context.
Also replaced the instance variable `self.mock_sha256` with a local variable in `setUp` as it is not used in other methods, and added `super().setUp()` to address a linter warning.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ie797499bd5475987ae6c4c38791ed92d38992a7c