[Refactor] Refactor get_iterations_summary in result_reporter.py
Refactor get_iterations_summary to use list join instead of string concatenation in a loop. This is generally considered more idiomatic and efficient in Python.
Also fix the return type hint of the method from None to str, as it returns a string.
These changes do not alter the existing logic or behavior of the code.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I8b475272af65e8c6ed57685fc370e5422f035c75