[Refactor] Use autospec=True and return_value for mock.patch
Replaced a lambda function used for mocking '_get_finder_instance_dict' with 'autospec=True' and 'return_value'. This improves readability and consistency with other mocks in the 'setUp' method, and makes the mock more robust. This change is purely an internal refactoring of the test setup and does not alter any existing logic or behavior of the code.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ib1c45d43bcb0afe0b8bb45ba79bd9eef23168891