[Refactor] doc: Add period to TestDetail.__init__ docstring
This change adds a period to the end of the docstring for the `__init__` method
of the `TestDetail` class. This improves consistency with other docstrings in
the file and addresses a g-short-docstring-punctuation lint warning.
The change is purely cosmetic and does not alter the logic or functionality
of the code.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ifc14ceab490379275356feb3fa2714891140b36f