[Refactor] Remove redundant mock.patch.stopall in test_finder_handler_unittest.py
The `mock.patch.stopall()` call in the `tearDown` method of
`TestFinderHandlerUnittests` is redundant. The `setUp` method already uses
`self.enterContext(mock.patch(...))` for all its mocks, which automatically
handles the cleanup of patches when the test method finishes or when `tearDown`
is called. Removing this redundant call improves clarity and removes unnecessary
code without altering the existing logic or behavior of the tests.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: If4b944847dd5672a584b8cd9988123fc9d2205ba