[Refactor] Introduce _COMMON_PACKAGE_REF_TYPES constant
This change introduces a new constant  to replace duplicated inline lists of reference types in . This improves readability and reduces repetition within the test file. The logic and functionality of the code remain unchanged, as it only replaces literal list definitions with a named constant.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I9f0630d7ee817c1b5d0bec685e9aa2396e88d17c