[Refactor] Use raw string for TestDetail.__init__ docstring
This change updates the docstring of the TestDetail.__init__ method to use a raw string.
This addresses a gpylint warning (C6118) regarding backslashes in docstrings, improving adherence to Python style guidelines.
This is a purely stylistic change and does not alter the existing logic or behavior of the code in any way.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I636463df298018f51d484c09b9e6a853a6ad3151