[Refactor] Simplify f-string in TestDetail.__str__
This change simplifies the f-string used to format the options string in the TestDetail.__str__ method. Instead of using a list comprehension within the join method, it directly uses a generator expression.

This improves readability and conciseness without altering the existing logic or functionality of the code.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ifb29329e301b85c1da7d573b49dd495eec0c8758