[Refactor] Improve readability in test_mapping.py by using 'in' and 'any()'
This change refactors the 'is_match_file_patterns' function in test_mapping.py to enhance readability and conciseness. It replaces a nested loop structure with Pythonic 'in' operator and 'any()' calls.

Specifically, the check for 'constants.TEST_MAPPING' is now a direct 'in' check, which is clearer. The subsequent pattern matching is also refactored using 'any()' with a generator expression. This makes the code easier to understand by expressing the intent more directly without altering the original logic or behavior of the code.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I36149fe81597091764915b6f47d18e7f5d02fe50