[Refactor] Improve assert message in TestDetail.__init__
This change enhances the clarity of the assertion message in the TestDetail constructor. Previously, if the 'host' parameter was not a boolean, the error message was generic. Now, it explicitly includes the type and value of the problematic 'host' parameter.

This refactoring improves debuggability without altering the existing logic or behavior of the code, as it only affects the error message during an assertion failure. The functionality remains identical when the 'host' parameter is valid.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: Ibe567023f8d82508b6c9b21a8e9fd78837b00d0c