[Refactor] Remove duplicate code in get_path of test_mapping.py
The get_path method in test_mapping.py contained duplicated logic for checking the existence of a path relative to different base directories. This refactor introduces a loop to iterate over the base directories, simplifying the code and adhering to the DRY principle.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I9b5cd910ca38bf0af4367ff478bcc5016a9f60a5