[Refactor] Combine nested if statements in _validate_ref
Combine the nested `if` statements in `_validate_ref` into a single conditional. This improves readability by reducing indentation and making the condition clearer. It does not alter existing logic as both conditions must be true for the body to execute, which matches the behavior of `and`.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

Change-Id: I17c77b8576ef4c79f6ec9cb383945b9621641324